splitting out PropagatorOptions.hpp
